LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/layer-cssom-order-reverse-expected.txt

Lorem ipsum
Lorem ipsum

PASS Insert layer invalidates style
PASS Delete layer invalidates style
PASS Insert layer invalidates @font-face
PASS Delete layer invalidates @font-face
FAIL Insert layer invalidates @property assert_equals: expected "green" but got ""
FAIL Delete layer invalidates @property assert_equals: expected "green" but got ""
